+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v21.1.3 (Unreleased)
 ### 🛠 Bug fixes
 * Renew token with the Vault Agent where the Key Manager is watching "VAULT_TOKEN_FILE". The Key Manager supports plaintext token and wrapped-token
+* Metric value for job status update CREATED to STARTED 
 
 ## v21.1.2 (2021-02-25)
 ### 🆕 Features